Adds helper repl init-ns, re-organizes src files
I added a helper namespace, `soundlab.user` which is set as the REPL
init-ns, which contains a helper function `ce` for quickly connecting
to an external server. The plan is for this namespace to contain other
helper functions for controlling the Overtone environment. For
example, when using this in a normal REPL I often stop the sound,
reload the ns and then start the sound... every time I make a change.
This could be a helper function. I also re-organized the source code
files so that core.clj contains instruments and generic composition
functions and then my individual drawing boards (random experiments)
are named by date e.g. 20210415.clj. These can be renamed to something
more memorable if desired.