Updates README; fixes bugs in soundlab.core; adds new composition
I updated the README to include instructions for `pasuspender` instead
of stopping pulseaudio entirely. In the README, I also increased the
SuperCollider server memory because I was getting an error saying that
I was running out of memory. I fixed some simple bugs with a previous
refactor in soundlab.core. I added a new dependency - leipzig. It's
supposed to help with composition. I added a new sketch which uses it
and the result was great. It was easy to use and fun to work with, so
I plan to leverage that more.